refactor: added components folder and put all components there, changed X to IconX
