+++ apps/web/app/assets/svgs/index.ts
@@ -41,7 +41,7 @@ export { default as FreeRight } from "./free-right.svg?react";
 export { default as NoData } from "./no-data.svg?react";
 export { default as Eye } from "./eye.svg?react";
 export { default as Enroll } from "./enroll.svg?react";
-export { default as X } from "./x.svg?react";
+export { default as IconX } from "./x.svg?react";
 export * from "./lesson-types";
 export * from "./question-types";
 export * from "./arrows";
