+++ apps/web/app/modules/Admin/Groups/EditGroup.page.tsx
@@ -8,10 +8,10 @@ import { useGroupByIdQuerySuspense } from "~/api/queries/admin/useGroupById";
 import { GROUPS_QUERY_KEY } from "~/api/queries/admin/useGroups";
 import { queryClient } from "~/api/queryClient";
 import { PageWrapper } from "~/components/PageWrapper";
-import CreateGroupCard from "~/modules/Admin/Groups/CreateGroupCard";
+import CreateGroupCard from "~/modules/Admin/Groups/components/CreateGroupCard";
+import { GroupPageBreadcrumbs } from "~/modules/Admin/Groups/components/GroupBreadcrumbs";
+import { GroupHeader } from "~/modules/Admin/Groups/components/GroupHeader";
 import { groupFormSchema } from "~/modules/Admin/Groups/group.utils";
-import { GroupPageBreadcrumbs } from "~/modules/Admin/Groups/GroupBreadcrumbs";
-import { GroupHeader } from "~/modules/Admin/Groups/GroupHeader";
 import Loader from "~/modules/common/Loader/Loader";
 
 import type { ReactElement } from "react";
