+++ apps/web/app/modules/Admin/Groups/components/RequiredTick.tsx
@@ -1,0 +1,3 @@
+export const RequiredTick = () => {
+  return <span className="text-error-600">*&nbsp;</span>;
+};
